+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+Added:
+
+- Add Open Settings command; add
+- Add Sketch Explorer menu w/ Create Sketch
+
 ## [1.1.3] - 2021-10-03
 
 Fixed: